WIP: switching to MVC paradigm in Polymer UI
SparkPolymerUI receives a handle to SparkModel, which Spark and SparkPolymer implement. SparkPolymerUI uses the model to map various events in the UI to ActionManager calls, etc.

There is currently a race condition: Spark ctor wants to access various elements of the UI, and so the UI (SparkPolymerUI) gets created too early, before enteredView() fires for various components, leading to a crash in SparkSelector on a null _target.