Temporarily patched the MVC scheme
Failed to use a full-on MVC with programmatic creation of SparkPolymerUI from SparkPolymer ctor: that's too late, because the super-ctor (Spark) queries the still-uncreated UI via querySelector, and thus crashes. After a switchover to Polymer and getting rid of querySelector's to do things, the full MVC will be possible to implement.