Fix the deployed UI (almost everything works)
Both undeployed and deployed UIs look fine. However, the menu button in the deployed UI doesn't initially show up; when hovered over and clicked, it does show up, but doesn't expand the menu.

The core of the fix is switching from @host to :host in the elements' CSS for the deployed UI - this was thge reason the styles weren't applied.

However, Dartium is still using @host, so it's temporarily left in, too.