Fixed the root cause of ghost dialogs lingering after closing and blocking UI
The problem affected <spark-modal>, which extends <spark-overlay>. <spark-overlay> relies on automaticaly generated Webkit events, such as on-webkitTransitionEnd, to complete its lifecycle. It defines these handlers in its HTML definition. It turns out that such event handlers are not inherited by extending class, so the events were not delivered to instances of <spark-modal>, breaking their lifecycle. A solution was to redeclare the handlers in <spark-modal> definition.