Fix sticky menu selection
Selection emained after a menu item click and was shown on the next menu opening.

The main reason is a bug in Chrome that fails to remove the :hover style from an element when that element is clicked and programmatically moved out from under mouse in some way. In our case, the menu auto-closes on a click, so the item remained in the :hover state. A solution was to use own CSS class and switch it on/off on mouseover/mouseout events.