Add support for pushing a Chrome app to the Chrome ADT on mobile.
This adds a new button next to "Run" that packages a CRX file and sends
it to a mobile device. The protocol for pushes is defined at:
https://github.com/MobileChromeApps/harness-push

Caveat: This code uses chrome.socket to make the HTTP request directly,
due to problems encountered while trying to use HttpRequest. The
HttpRequest code is much more compact, if the issues with it are
resolved.

Caveat: It currently pushes all of the first project in the workspace.
Once the upcoming work to be much more context-aware about what app
we're launching or pushing lands, this should be updated.