Switched from 'autofocus' to 'focused' to designate autofocused elements
'autofocus' is a W3C standard, so an attempt to use it must have clashed with the browser handling of it. A number of weird things started to happen, e.g. the menu instead of fully closing started to look like visibility:hidden, i.e. the elements' bounding boxes were visible in Dev Tools and received clicks.