Improve usability of cannot-launch alert.
The previous language didn't indicate which way the flag should be
set, but merely said to "check" it. The new language is more
precise. The alert now presents a link that can be manually copied
and pasted into an address bar, taking the user directly to the
relevant flag.

TBR @sergeygs