Deploy via USB can now communicate with a local ADB server
If the ADB server is already running, Spark will use that to deploy to
mobile instead of its custom USB code. This avoids fighting with ADB
over the USB interface.

Chrome DevTools likewise uses the system's ADB server if it exists, so
when the ADB server is running Spark and DevTools coexist peacefully.

On Chromebooks, nothing has changed.