Add package-files-are-editable dev flag
This is to facilitate debugging of package-related issues.

One thing that doesn't work currently: if a package file is open while the flag is off, and then the flag is turned on and CDE is restarted, the reloaded source is still read-only. It needs to be closed and reopened to make it writable.