Adjust template to Polymer/platform having moved to Polymer/webcomponents #2
It turns out Polymer is in a bit of flux right now. The switch to webcomponents is incomplete: webcomponents itself is not officially released yet, so paper elements still depend on platform, but polymer no longer includes platform, so it needs to be imported explicitly in bower.json.

After the switch is complete, this PR should be undone to the way the code was right before it.